+++ app.h
@@ -121,7 +121,7 @@ std::string otio_error_string(otio::ErrorStatus const& error_status);
 ImU32 ImLerpColors(ImU32 col_a, ImU32 col_b, float t);
 
 void SelectObject(otio::SerializableObject* object, otio::SerializableObject* context=NULL);
-void SeekPlayhead(float seconds);
+void SeekPlayhead(double seconds);
 void SnapPlayhead();
 void FitZoomWholeTimeline();
